examples/tensorflow-profiling-examples/example.py
from absl import logging

import tensorflow as tf


@tf.keras.utils.register_keras_serializable(package='Custom')
class LinearBlockFull(tf.keras.layers.Layer):

    def __init__(self, units=32, **kwargs):
        super(LinearBlockFull, self).__init__(**kwargs)
        self.units = units

    def build(self, input_shape):
        self.w = self.add_weight(shape=(input_shape[-1], self.units),
                                 initializer='random_normal',
                                 trainable=True)
        self.b = self.add_weight(shape=(self.units,),
                                 initializer='zeros',
                                 trainable=True)

    def call(self, inputs):
        return tf.matmul(inputs, self.w) + self.b

    def get_config(self):
        config = super(LinearBlockFull, self).get_config()
        custom_config = {'units': self.units}
        config.update(custom_config)
        return config


class LinearBlock(tf.keras.layers.Layer):

    def __init__(self, units=32):
        super(LinearBlock, self).__init__()
        self.units = units

    def build(self, input_shape):
        self.w = self.add_weight(shape=(input_shape[-1], self.units),
                                 initializer='random_normal',
                                 trainable=True)
        self.b = self.add_weight(shape=(self.units,),
                                 initializer='zeros',
                                 trainable=True)

    def call(self, inputs):
        return tf.matmul(inputs, self.w) + self.b


def get_model(dim):
    model = tf.keras.Sequential([
        tf.keras.layers.Dense(64, activation='relu', input_shape=[dim]),
        tf.keras.layers.Dense(32, activation='relu'),
        tf.keras.layers.Dense(1)
    ])
    model.summary(print_fn=logging.info)
    optimizer = tf.keras.optimizers.RMSprop(0.001)
    model.compile(loss='mse',
                  optimizer=optimizer,
                  metrics=['mae', 'mse'])
    return model
